moved templates into sub-folder per the new ADR
